Implemented Lever Functionality
Implemented functionality for the lever that can be used for either the aimAxis or shotPower.
The ControllStick will pass a float between 0 and 100 to the CatapultControl based on the position of the Lever.
I thought it would probably be best if interoperating this value was handled in the CatapultControl so it would be easier to change.